+++ API/Controllers/BookController.cs
@@ -2,11 +2,13 @@
 using Core.Enums;
 using Core.Exceptions;
 using Core.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers
 {
+    [Authorize]
     [EnableCors("CorsPolicy")]
     [Route("libraryApi/[controller]")]
     [ApiController]
@@ -21,11 +23,14 @@ namespace API.Controllers
         #endregion
 
         #region Get
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("All")]
         public async Task<IActionResult> GetAllAsync()
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var books = await _bookService.GetAllAsync();
                 return Ok(new { message = "ok", response = books });
             }
@@ -35,11 +40,14 @@ namespace API.Controllers
             }
         }
 
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("AllWithIncludes")]
         public async Task<IActionResult> GetAllWithIncludesAsync()
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var books = await _bookService.GetAllWithIncludesAsync();
                 return Ok(new { message = "ok", response = books });
             }
@@ -49,11 +57,14 @@ namespace API.Controllers
             }
         }
 
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("Specification")]
         public async Task<IActionResult> GetBySpecificationAsync(string? title, Genre? genre, decimal? minPrice, decimal? maxPrice)
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var books = await _bookService.GetBySpecificationAsync(title, genre, minPrice, maxPrice);
                 return Ok(new { message = "ok", response = books });
             }
@@ -63,11 +74,14 @@ namespace API.Controllers
             }
         }
 
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("SpecificationWithIncludes")]
         public async Task<IActionResult> GetBySpecificationWithIncludesAsync(string? title, Genre? genre, decimal? minPrice, decimal? maxPrice)
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var books = await _bookService.GetBySpecificationWithIncludesAsync(title, genre, minPrice, maxPrice);
                 return Ok(new { message = "ok", response = books });
             }
@@ -77,11 +91,14 @@ namespace API.Controllers
             }
         }
 
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetByIdAsync(int id)
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var book = await _bookService.GetByIdAsync(id);
                 return Ok(new { message = "ok", response = book });
             }
@@ -95,11 +112,14 @@ namespace API.Controllers
             }
         }
 
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks, ViewAuthorsBooks")]
         [HttpGet("WithIncludes/{id:int}")]
         public async Task<IActionResult> GetByIdWithIncludesAsync(int id)
         {
             try
             {
+                if (User.HasClaim("BooksAccess", "false")) return Forbid();
+
                 var book = await _bookService.GetByIdWithIncludesAsync(id);
                 return Ok(new { message = "ok", response = book });
             }
@@ -115,11 +135,14 @@ namespace API.Controllers
         #endregion
 
         #region Post
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks")]
         [HttpPost("Add")]
         public async Task<IActionResult> AddAsync([FromBody] BookDTO bookDTO)
         {
             try
             {
+                if (User.HasClaim("BooksCreate", "false")) return Forbid();
+
                 if (!ModelState.IsValid) return BadRequest(ModelState);
 
                 await _bookService.AddAsync(bookDTO);
@@ -137,11 +160,14 @@ namespace API.Controllers
         #endregion
 
         #region Put
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks")]
         [HttpPut("Update")]
         public async Task<IActionResult> UpdateAsync([FromBody] BookDTO bookDTO)
         {
             try
             {
+                if (User.HasClaim("BooksEdit", "false")) return Forbid();
+
                 if (!ModelState.IsValid) return BadRequest(ModelState);
 
                 await _bookService.UpdateAsync(bookDTO);
@@ -159,11 +185,14 @@ namespace API.Controllers
         #endregion
 
         #region Delete
+        [Authorize(Roles = "Admin, AuthorsBooksPublisher, AuthorsBooks")]
         [HttpDelete("Delete/{id:int}")]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             try
             {
+                if (User.HasClaim("BooksDelete", "false")) return Forbid();
+
                 await _bookService.DeleteAsync(id);
                 return Ok(new { message = "ok" });
             }
